d3/requirements.txt
opencv-python-headless
numpy
apache-beam[gcp]
google-cloud-pubsub
google-cloud-storage
